Adding support for multiple phis  - i.e. multiple influence functions
Note the phis file is just a collection of functions, which it is assumed are 1-indexed (and in particular written in order of indexing).
Support could be added to not assume this ordering by converting the list of tuples that is produced by inspect to a dict, but it is convenient right now.
It will likely be changed soon.